refactor: create module for APIs requests
