+++ api/quizzify/api/__init__.py
@@ -1,0 +1,1 @@
+"""Module for API requests to Spotify and LastFM."""
