+++ api/quizzify/api/spotify/spotify_requests.py
@@ -334,46 +334,3 @@ def spotify_get_artist_info_from_name(artist_name: str):
             status_code=response.status_code,
             detail="Failed to retrieve artist information",
         )
-
-
-def spotify_get_related_artists(artist_id: str):
-    """Get the related artists for an artist from Spotify.
-
-    Parameters
-    ----------
-    artist_id : str
-        The Spotify ID for the artist.
-
-    Returns
-    -------
-    list
-        A list of the related artists.
-    """
-    headers = spotify_headers()
-    api_url = f"{SPOTIFY_BASE_URL}/artists/{artist_id}/related-artists"
-    response = requests.get(
-        api_url,
-        headers=headers,
-        timeout=120,
-    )
-
-    if response.status_code == 200:
-        raw_related_artists = response.json()["artists"]
-        related_artists = []
-        for raw_artist in raw_related_artists:
-            best_image = get_highest_resolution_image(images=raw_artist["images"])
-            current_artist = {
-                "id": raw_artist["id"],
-                "name": raw_artist["name"],
-                "popularity": raw_artist["popularity"],
-                "genres": raw_artist["genres"],
-                "followers": raw_artist["followers"]["total"],
-                "image_url": best_image["url"] if best_image else None,
-            }
-            related_artists.append(current_artist)
-        return related_artists
-    else:
-        raise HTTPException(
-            status_code=response.status_code,
-            detail=f"Failed to retrieve related artists for '{artist_id}'",
-        )
