+++ api/quizzify/question/song/question_song_artist.py
@@ -1,11 +1,11 @@
 import logging
 import random
 
+from quizzify.api.lastfm.lastfm_requests import lastfm_get_similar_artists
+from quizzify.api.spotify.spotify_requests import spotify_get_artist_info_from_name
 from quizzify.crud import artists as crud_artists
-from quizzify.lastfm.lastfm_requests import lastfm_get_similar_artists
 from quizzify.question.abstract_question import AbstractQuestion
 from quizzify.question.question_types import SongQuestionType
-from quizzify.spotify.spotify_requests import spotify_get_artist_info_from_name
 from quizzify.utils.schemas import Artist
 
 logger = logging.getLogger(__name__)
