+++ api/quizzify/routers/songs/router.py
@@ -2,8 +2,8 @@ from typing import Dict, List
 
 from fastapi import APIRouter
 
+from quizzify.api.spotify.spotify_requests import spotify_get_user_id
 from quizzify.routers.songs import service
-from quizzify.spotify.spotify_requests import spotify_get_user_id
 from quizzify.utils.schemas import TimeRange
 
 # define router for songs endpoints
